test(eval): fold config-bundle run golden into ab-test.fixture.test.tsx
